fix: ValueError: parquet doesn't support non-string colum names
yf.download() returns a MultiIndex in the columns  attribute of the dataframe (because you can use it to download data from multiple tickets at once), while history() returns a simple Index. So when the code in the save_by_year(...) method of your DataManager class tries to save the dataframe as a Parquet file, the ValueError exception is thrown.